requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
fake-useragent>=1.4.0
pandas>=2.0.0
ibmcloudant>=0.8.0
python-dotenv>=1.0.0
